+++ src/frontend/apps/impress/src/core/index.ts
@@ -1,4 +1,3 @@
 export * from './AppProvider';
 export * from './auth';
-export * from './conf';
 export * from './config';
